+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.6.0] - 2018-07-03
+
+### Added
+
+- `scoped` interactor method to return nested and scoped interactors
+- ability to specify a selector function for collections
+
 ## [0.5.1] - 2018-07-02
 
 ### Fixed